requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
openai==1.58.1
gitpython==3.1.44
jinja2==3.1.5
pyyaml==6.0.2
markdown==3.7
apscheduler==3.10.4
python-dotenv==1.0.1
